Creates testing folder.
