+++ web_console_v2/api/testing/common.py
@@ -18,10 +18,12 @@ import json
 import logging
 import secrets
 from http import HTTPStatus
+from flask import Flask
 from flask_testing import TestCase
 from fedlearner_webconsole.app import create_app
 from fedlearner_webconsole.db import db
 from fedlearner_webconsole.auth.models import User
+
 
 class BaseTestCase(TestCase):
     SQLALCHEMY_DATABASE_URI = 'sqlite://'
@@ -109,3 +111,13 @@ class BaseTestCase(TestCase):
         return resp
 
 
+def create_test_db():
+    """Creates test db for testing non flask-must units."""
+    app = Flask('fedlearner_webconsole_test')
+    app.config['TESTING'] = True
+    app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///:memory:'
+    app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
+    db.init_app(app)
+    # this does the binding
+    app.app_context().push()
+    return db
